no project exception, add tests
Made it so there is an exception when no project is found with `utils.find_project`. Additionally, added `test_perforrm_parsimony` to the tests in project_initialization. Will most likely have to check this a couple of times.

NOTE: The project created is not deleted, so I can look at which models are kept.